Rename modules to imc_modules and update imports
