+++ README.md
@@ -1,5 +1,6 @@
 # imcsdk_test
 
 Example modules demonstrating features of Cisco IMC SDK and UCS Manager SDK.
-See `docs/` for API references for both SDKs.
+IMC-related helpers live under `imc_modules/` and UCS Manager helpers under
+`ucs_modules/`. See `docs/` for API references for both SDKs.
