+++ imc_modules/manual_discovery.py
