+++ ucs_modules/automatic_discovery.py
@@ -1,6 +1,6 @@
 """Automatic discovery of UCS Manager servers in a given IP range."""
 from ucsmsdk.ucshandle import UcsHandle
-from modules.utils import validate_sdk_method
+from imc_modules.utils import validate_sdk_method
 
 login_file, login_line = validate_sdk_method(UcsHandle.login)
