+++ ucs_modules/backup.py
@@ -1,6 +1,6 @@
 """Backup utilities using ucsmsdk."""
 from ucsmsdk.utils import ucsbackup
-from modules.utils import validate_sdk_method
+from imc_modules.utils import validate_sdk_method
 
 backup_file, backup_line = validate_sdk_method(ucsbackup.backup_ucs)
