+++ ucs_modules/bmc_settings.py
@@ -1,6 +1,6 @@
 """BMC network configuration utilities for UCS."""
 from ucsmsdk.ucshandle import UcsHandle
-from modules.utils import validate_sdk_method
+from imc_modules.utils import validate_sdk_method
 
 query_file, query_line = validate_sdk_method(UcsHandle.query_classid)
