+++ ucs_modules/management.py
@@ -1,7 +1,7 @@
 """UCS management utilities."""
 from ucsmsdk.ucshandle import UcsHandle
 from ucsmsdk.mometa.org.OrgOrg import OrgOrg
-from modules.utils import validate_sdk_method
+from imc_modules.utils import validate_sdk_method
 
 add_file, add_line = validate_sdk_method(UcsHandle.add_mo)
 commit_file, commit_line = validate_sdk_method(UcsHandle.commit)
