+++ ucs_modules/manual_discovery.py
@@ -1,6 +1,6 @@
 """Manual discovery helper for a single UCS target."""
 from ucsmsdk.ucshandle import UcsHandle
-from modules.utils import validate_sdk_method
+from imc_modules.utils import validate_sdk_method
 
 login_file, login_line = validate_sdk_method(UcsHandle.login)
